Module numbers_rus::double::double_vector::double_vector

Functions

two_vector_median

Returns the median of two vectors as a string

two_vector_median_float

Returns the median of two float vectors as a string

two_vector_mode

Returns the mode of two combined integer vectors as a string

two_vector_mode_float

vector_add

Returns the sum of two vectors as a string

vector_add_float

Returns the sum of two float vectors as a string

vector_average

Returns the average of two vectors as a string

vector_average_float

Returns the average of two float vectors as a string

vector_divide

Returns the quotient of two vectors as a string

vector_divide_float

Returns the quotient of two float vectors as a string

vector_max

Returns the maximum of two vectors as a string

vector_max_float

Returns the maximum of two float vectors as a string

vector_min

Returns the minimum of two vectors as a string

vector_min_float

Returns the minimum of two float vectors as a string

vector_modulo

Returns the remainder of two vectors as a string

vector_modulo_float

Returns the modulo of two float vectors as a string

vector_multiply

Returns the product of two vectors as a string

vector_multiply_float

Returns the product of two float vectors as a string

vector_power

Returns the power of two vectors as a string

vector_power_float

Returns the power of two float vectors as a string

vector_root

Returns the root of two vectors as a string

vector_root_float

Returns the root of two float vectors as a string

vector_subtract

Returns the difference of two vectors as a string

vector_subtract_float

Returns the difference of two float vectors as a string