Module numbers_rus::floats::base_float::base_float

Functions

add_float

This function takes two f64 numbers as arguments, left and right, and returns their sum as an f64. The calculation is performed using native

divide_float

This function takes two f64 numbers as arguments, left and right, and returns their quotient as an f64. The calculation is performed using native

factorial_float

Calculates the factorial of a given f64 number. Returns the factorial result as an f64. It uses recursion to calculate the factorial.

fibonacci_float

Calculates the Fibonacci number at a given f64 index, using recursion. Returns the Fibonacci result as f64.

is_even_float

Checks whether the given f64 number is even. Returns true if the number is even, otherwise false. The function calculates the remainder of the input number divided by 2.0 and compares it to 0.0. If they are equal, the input number is even.

is_odd_float

Checks whether the given f64 number is odd. Returns true if the number is odd, otherwise false. The function calculates the remainder of the input number divided by 2.0 and compares it to 0.0. If they are not equal, the input number is odd.

is_perfect_cube_float

Determines whether number is a perfect cube by comparing its cube root cubed with the original number. Returns true if the number is a perfect cube, otherwise false.

is_perfect_power_float

Determines if the given f64 number is a perfect power by comparing its square root squared with the original number. Returns true if the number is a perfect power, otherwise false.

is_perfect_square_float

Determines whether number is a perfect square by comparing its square root squared with the original number. Returns true if the number is a perfect square, otherwise false.

is_prime_float

Returns true if the given f64 number is prime. This function checks for primality by iterating through numbers from 2 to the square root of the given number

modulo_float

This function takes two f64 numbers as arguments, left and right, and returns the result of the modulo operation (left modulo right) as an f64. The calculation is performed using the native modulo operator.

multiply_float

This function takes two f64 numbers as arguments, left and right, and returns their product as an f64. The calculation is performed using native

power_float

This function takes two f64 numbers as arguments, left and right, and returns the left raised to the power of right as an f64. The calculation is performed using the native powf method.

root_float

This function takes two f64 numbers as arguments: left and right, and returns the left raised to the power of 1 / right as an f64. The calculation is performed using the native powf method.

subtract_float

This function takes two f64 numbers as arguments, left and right, and returns their difference as an f64 where left is substracted by right.