Module numbers_rus::single::single_vector::single_vector

Functions

vector_interquartile_range

Computes the interquartile range (IQR) of a given vector of 128-bit signed integers.

vector_interquartile_range_float

Computes the interquartile range (IQR) of a given vector of 64-bit floating-point numbers.

vector_mean

Calculates the mean of a given vector of 128-bit signed integers.

vector_mean_float

Calculates the mean of a given vector of 64-bit floating-point numbers.

vector_median

Calculates the median value of a vector containing 128-bit signed integers.

vector_median_float

Calculates the median value of a given vector of 64-bit floating-point numbers.

vector_mode

vector_mode_float

vector_product

Returns the product of all elements in a vector of 128-bit signed integers.

vector_product_float

Computes the product of all elements in a given vector of floating-point numbers.

vector_quartiles

Computes the first (Q1), second (Q2), and third (Q3) quartiles of a given vector of 128-bit signed integers.

vector_quartiles_float

Returns a formatted string containing the first (Q1), second (Q2), and third (Q3) quartiles of a given vector of 64-bit floating-point numbers.

vector_range

Calculates the range of a given vector of 128-bit signed integers, which is the difference between the maximum and minimum values in the vector.

vector_range_float

Calculates the range of a given vector of 64-bit floating-point numbers, which is the difference between the maximum and minimum values in the vector.

vector_standard_deviation

Calculates the standard deviation of a given vector of 128-bit signed integers.

vector_standard_deviation_float

Calculates the standard deviation of a given vector of 64-bit floating-point numbers.

vector_sum

Returns the sum of all elements in a vector of 128-bit signed integers.

vector_sum_float

Computes the sum of all elements in a given vector of floating-point numbers.

vector_variance

Calculates the variance of a given vector of 128-bit signed integers.

vector_variance_float

Calculates the variance of a given vector of 64-bit floating-point numbers.